requirements.txt
flask
flask-socketio
redis
python-dotenv
paho-mqtt
aiomqtt
boto3